Refactor rendering pipeline and enhance picking functionality
- Updated RenderSceneController to emit scene update signals instead of camera change signals.
- Introduced SceneRendererImpl to encapsulate the rendering pipeline, including geometry, mesh, highlight, and picking passes.
- Enhanced picking logic to support pixel-level hit detection with improved entity resolution.
- Refactored RenderBatch to streamline mesh uploads and drawing operations, including new methods for drawing indexed and vertex ranges.
- Updated shaders to accommodate new UID handling for picking.
- Modified SelectManager to work with RenderEntityType and updated selection methods accordingly.
- Improved overall code organization and clarity, including better encapsulation of rendering logic and helper functions.

+++ include/app/select_manager_service.hpp
@@ -59,14 +59,14 @@ public:
      * @param entity_uid Type-scoped entity id
      * @param entity_type Entity type name (e.g. "Edge", "Face")
      */
-    Q_INVOKABLE void selectEntity(uint32_t entity_uid, const QString& entity_type);
+    Q_INVOKABLE void selectEntity(uint64_t entity_uid, const QString& entity_type);
 
     /**
      * @brief Remove an entity from the current selection by uid and type
      * @param entity_uid Type-scoped entity id
      * @param entity_type Entity type name (e.g. "Edge", "Face")
      */
-    Q_INVOKABLE void removeEntity(uint32_t entity_uid, const QString& entity_type);
+    Q_INVOKABLE void removeEntity(uint64_t entity_uid, const QString& entity_type);
 
     /**
      * @brief Check whether an entity (uid+type) is selected
@@ -74,18 +74,18 @@ public:
      * @param entity_type Entity type name
      * @return true if the entity is in the current selection set
      */
-    Q_INVOKABLE bool isEntitySelected(uint32_t entity_uid, const QString& entity_type) const;
+    Q_INVOKABLE bool isEntitySelected(uint64_t entity_uid, const QString& entity_type) const;
 
     /**
      * @brief Get all current selections
      * @return Vector of (uid, type) pairs
      */
-    Q_INVOKABLE QVector<QPair<uint32_t, QString>> currentSelections() const;
+    Q_INVOKABLE QVector<QPair<uint64_t, QString>> currentSelections() const;
 
 signals:
     void selectModeChanged(uint32_t select_types);
-    void entitySelected(uint32_t entity_uid, const QString& entity_type);
-    void entityRemoved(uint32_t entity_uid, const QString& entity_type);
+    void entitySelected(uint64_t entity_uid, const QString& entity_type);
+    void entityRemoved(uint64_t entity_uid, const QString& entity_type);
     void selectModeActivated(bool enabled);
     void selectionCleared();
